update Matter 1.3

+++ scripts/setup/bootstrap.sh
@@ -21,14 +21,14 @@ _install_additional_pip_requirements() {
     # figure out additional pip install items
     while [ $# -gt 0 ]; do
         case $1 in
-            -p | --platform)
-                _SETUP_PLATFORM=$2
-                shift # argument
-                shift # value
-                ;;
-            *)
-                shift
-                ;;
+        -p | --platform)
+            _SETUP_PLATFORM=$2
+            shift # argument
+            shift # value
+            ;;
+        *)
+            shift
+            ;;
         esac
     done
 
@@ -41,7 +41,7 @@ _install_additional_pip_requirements() {
 
         for platform in ${_SETUP_PLATFORM}; do
             # Allow none as an alias of nothing extra installed (like -p none)
-            if [ "$platform" != "none" ]; then
+            if [ "$platform" != "none" -a -e "$_CHIP_ROOT/scripts/setup/requirements.$platform.txt" ]; then
                 echo "Installing pip requirements for $platform..."
                 pip install -q \
                     -r "$_CHIP_ROOT/scripts/setup/requirements.$platform.txt" \
@@ -66,23 +66,13 @@ _bootstrap_or_activate() {
     local _BOOTSTRAP_NAME="${_BOOTSTRAP_PATH##*/}"
     local _BOOTSTRAP_DIR="${_BOOTSTRAP_PATH%/*}"
     # Strip off the 'scripts[/setup]' directory, leaving the root of the repo.
-    _CHIP_ROOT="$(cd "${_BOOTSTRAP_DIR%/setup}/.." && pwd)"
+    _CHIP_ROOT="$(cd "${_BOOTSTRAP_DIR%/setup}/.." >/dev/null && pwd)"
 
     local _CONFIG_FILE="scripts/setup/environment.json"
 
     if [ -n "$PW_CONFIG_FILE" ]; then
         _CONFIG_FILE="$PW_CONFIG_FILE"
         unset PW_CONFIG_FILE
-    fi
-
-    if [ ! -f "$_CHIP_ROOT/third_party/pigweed/repo/pw_env_setup/util.sh" ]; then
-        # Make sure our submodule remotes are correct for this revision.
-        git submodule sync --recursive
-        git submodule update --init
-    elif [ "$_BOOTSTRAP_NAME" = "bootstrap.sh" ]; then
-        # In this case, only update already checked out submodules.
-        git submodule sync --recursive
-        git submodule update
     fi
 
     PW_BRANDING_BANNER="$_CHIP_ROOT/scripts/setup/banner.txt"
@@ -94,6 +84,10 @@ _bootstrap_or_activate() {
     PW_ROOT="$_CHIP_ROOT/third_party/pigweed/repo"
     export PW_ROOT
 
+    # Update or init the pigweed submodule if necessary. Don't touch any other submodules.
+    if [ "$_BOOTSTRAP_NAME" = "bootstrap.sh" -a ! -f "$_CHIP_ROOT/third_party/pigweed/repo/pw_env_setup/util.sh" ]; then
+        git submodule update --init "$_CHIP_ROOT/third_party/pigweed/repo"
+    fi
     . "$_CHIP_ROOT/third_party/pigweed/repo/pw_env_setup/util.sh"
 
     _chip_bootstrap_banner() {
@@ -105,7 +99,7 @@ _bootstrap_or_activate() {
 
     local _PW_BANNER_FUNC="_chip_bootstrap_banner"
 
-    # Force the Pigweed environment directory to be '.environment'
+    # Default the Pigweed environment directory to be '.environment'
     if [ -z "$PW_ENVIRONMENT_ROOT" ]; then
         export PW_ENVIRONMENT_ROOT="$PW_PROJECT_ROOT/.environment"
     fi
@@ -116,8 +110,12 @@ _bootstrap_or_activate() {
     export PW_DOCTOR_SKIP_CIPD_CHECKS=1
     export PATH # https://bugs.chromium.org/p/pigweed/issues/detail?id=281
 
+    local _PIGWEED_CIPD_JSON="$_CHIP_ROOT/third_party/pigweed/repo/pw_env_setup/py/pw_env_setup/cipd_setup/pigweed.json"
+    mkdir -p "$_PW_ACTUAL_ENVIRONMENT_ROOT"
+    local _GENERATED_PIGWEED_CIPD_JSON="$_PW_ACTUAL_ENVIRONMENT_ROOT/pigweed.json"
+    $_CHIP_ROOT/scripts/setup/gen_pigweed_cipd_json.py -i $_PIGWEED_CIPD_JSON -o $_GENERATED_PIGWEED_CIPD_JSON
+
     if test -n "$GITHUB_ACTION"; then
-        mkdir -p "$_PW_ACTUAL_ENVIRONMENT_ROOT"
         tee <<EOF >"${_PW_ACTUAL_ENVIRONMENT_ROOT}/pip.conf"
 [global]
 cache-dir = ${_PW_ACTUAL_ENVIRONMENT_ROOT}/pip-cache
@@ -131,7 +129,8 @@ EOF
         pw_bootstrap --shell-file "$_SETUP_SH" \
             --install-dir "$_PW_ACTUAL_ENVIRONMENT_ROOT" \
             --config-file "$_CHIP_ROOT/$_CONFIG_FILE" \
-            --virtualenv-gn-out-dir "$_PW_ACTUAL_ENVIRONMENT_ROOT/gn_out"
+            --virtualenv-gn-out-dir "$_PW_ACTUAL_ENVIRONMENT_ROOT/gn_out" \
+            --additional-cipd-file "$_GENERATED_PIGWEED_CIPD_JSON"
         pw_finalize bootstrap "$_SETUP_SH"
         _ACTION_TAKEN="bootstrap"
     else
